// src/core/settings/settings.interface.tsx
export interface ISimpleConfiguration extends IConfiguration {
    properties?: IConfiguratioProperties;
}
export interface IAdvanceConfiguration extends IConfiguration {

}

export type Configuration = ISimpleConfiguration | IAdvanceConfiguration;

export interface IPowerSettings {
    pauseOnChargerDisconnected: boolean;
    pauseOnLowBattery: boolean;
    resumeOnBatteryOk: boolean;
    resumeOnChargerConnected: boolean;
}

export interface IThermalSettings {
    pauseOnCPUTemperatureOverHeat: boolean;
    pauseOnCPUTemperatureOverHeatValue: number;
    resumeCPUTemperatureNormal: boolean;
    resumeCPUTemperatureNormalValue: number;
}
export interface ISettings {
    ready: boolean;
    uuid: string;
    configurations: Array<Configuration>;
    selectedConfiguration?: string;
    power: IPowerSettings;
    thermal: IThermalSettings;
    donation: number;
    printTime: number;
}